Added the second upgrade
Added p_upg2: more enhancements. Doubles base PPC, max lvl 5: formula: 50*2^lvl
