+++ game.html
@@ -10,11 +10,11 @@
     <div>
         <h2>Production</h2>
         <div id="buyButton" class="mainButton">
-            <button class="cookie" id="cookieUpgrade0" style="display: block; right: 0px;" onclick="addCookie()"><span>Click to make 1 point(s).</span></button>
+            <button class="cookie" id="cookieUpgrade0" style="display: block; right: 0px;" onclick="addPoint()"><span>Click to make 1 point(s).</span></button>
         </div>
         <div id="cookiesModal" class="statsModal">
             <p class="modalText">You have</p>
-            <p class="modalTextStat" id="cookiestext">0</p>
+            <p class="modalTextStat" id="pointstext">0</p>
             <p class="modalText">points</p>   
         </div>
     </div>
@@ -32,8 +32,8 @@
         <div id="clickUpgrade2">
             <button class="clickUpgrade" id="clickUpgrade2" onclick="upgrade2()">
                 <p class="clickUpgradeTitle">2. More enchancements</p>
-                <p class="clickUpgradeText">Double the number of points per click for</p>
-                <p class = "modalTextStat" id="upgrade2indicator">50</p>
+                <p class="clickUpgradeText">Base PPC x2</p>
+                <p class = "modalTextStat" id="upgrade2costindicator">50</p>
                 <p class="clickUpgradeText">points.</p>
             </button>
         </div>
